# api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mini RBAC Go API",
    "version": "2.0.0",
    "description": "Role-Based Access Control API for managing roles, groups, permissions, and workspace hierarchies",
    "contact": {
      "name": "Mini RBAC Go",
      "url": "https://github.com/redhat/mini-rbac-go"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8085/api/rbac/v2",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/groups": {
      "get": {
        "tags": ["Groups"],
        "summary": "List all groups",
        "description": "Returns a paginated list of groups for the tenant",
        "operationId": "listGroups",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Groups"],
        "summary": "Create a new group",
        "description": "Creates a new group",
        "operationId": "createGroup",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Group created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/groups/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/GroupId"
        }
      ],
      "get": {
        "tags": ["Groups"],
        "summary": "Get group by ID",
        "description": "Returns a single group",
        "operationId": "getGroup",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": ["Groups"],
        "summary": "Update group",
        "description": "Updates a group's name or description",
        "operationId": "updateGroup",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Group updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": ["Groups"],
        "summary": "Delete group",
        "description": "Deletes a group and all its associations",
        "operationId": "deleteGroup",
        "responses": {
          "204": {
            "description": "Group deleted successfully"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/groups/{id}/principals": {
      "parameters": [
        {
          "$ref": "#/components/parameters/GroupId"
        }
      ],
      "get": {
        "tags": ["Groups"],
        "summary": "List group principals",
        "description": "Returns all principals (users) in a group",
        "operationId": "listPrincipals",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrincipalListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "post": {
        "tags": ["Groups"],
        "summary": "Add principals to group",
        "description": "Adds users to a group",
        "operationId": "addPrincipals",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrincipalsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Principals added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      },
      "delete": {
        "tags": ["Groups"],
        "summary": "Remove principals from group",
        "description": "Removes users from a group",
        "operationId": "removePrincipals",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrincipalsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Principals removed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Group": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Group unique identifier"
          },
          "name": {
            "type": "string",
            "description": "Group name"
          },
          "description": {
            "type": "string",
            "description": "Group description"
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "modified": {
            "type": "string",
            "format": "date-time",
            "description": "Last modification timestamp"
          }
        }
      },
      "CreateGroupRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Group name",
            "example": "Engineering Team"
          },
          "description": {
            "type": "string",
            "description": "Group description",
            "example": "Engineering department group"
          }
        }
      },
      "UpdateGroupRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Updated group name"
          },
          "description": {
            "type": "string",
            "description": "Updated group description"
          }
        }
      },
      "GroupListResponse": {
        "type": "object",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Group"
            }
          }
        }
      },
      "PrincipalsRequest": {
        "type": "object",
        "required": ["principals"],
        "properties": {
          "principals": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of user IDs (email addresses)",
            "example": ["alice@example.com", "bob@example.com"]
          }
        }
      },
      "PrincipalResponse": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "description": "User ID (email address)",
            "example": "alice@example.com"
          }
        }
      },
      "PrincipalListResponse": {
        "type": "object",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PrincipalResponse"
            }
          }
        }
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "description": "Number of items in current response"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Error title"
          },
          "detail": {
            "type": "string",
            "description": "Detailed error message"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code"
          }
        }
      }
    },
    "parameters": {
...
      "GroupId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Group UUID",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  }
}